CreateUserForm moved to inside CreateUser
