+++ web-app/src/components/CreateUser/index.jsx
@@ -1,15 +1,15 @@
 import React from 'react';
 import { PropTypes as PT } from 'prop-types';
 import container from './container';
-import CreateUserForm from '../CreateUserForm';
-import { Card, Errorbox } from '../common';
+import CreateUserForm from './CreateUserForm';
+import { Errorbox } from '../common';
 
 export const CreateUser = props => {
   const { error, success, onSuccess, onFailed } = props;
 
   return (
-    <Card>
-      <h3>Create User</h3>
+    <div>
+      <h2>Create User</h2>
       <CreateUserForm onSuccess={onSuccess} onFailed={onFailed} />
       <Errorbox
         id="errorCreateUser"
@@ -17,7 +17,7 @@ export const CreateUser = props => {
         label="Error creating user"
       />
       {success && <p id="userCreatedSuccess">User created successfully!</p>}
-    </Card>
+    </div>
   );
 };
